add actions to header
